chore : price 모두 int형으로 통일

+++ src/main/kotlin/com/example/cakezip/domain/cake/CakeOptionList.kt
@@ -20,7 +20,7 @@ class CakeOptionList(
 
     var optionDetail: String,
 
-    var optionPrice: Long,
+    var optionPrice: Int,
 
     ) : BaseEntity() {
     @Id
